pyproject.toml
[project]
name = "cauliflow"
version = "0.1.0"
description = "cauliflow"
readme = "README.md"
authors = [{ name = "Shinya Sasaki", email = "shinya.sasaki@kek.jp" }]
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.8",
    "janus>=2.0.0",
    "lark>=1.2.2",
    "pyepics>=3.5.7",
    "pyyaml>=6.0.2",
    "zabbix-utils[async]>=2.0.2",
]

[project.scripts]
cauliflow = "cauliflow.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "tox>=4.26.0",
    "tox-uv>=1.25.0",
]

[tool.tox]
requires = ["tox>=4.19"]
env_list = ["3.13", "3.12", "3.11"]

[tool.tox.env_run_base]
description = "Run test under {base_python}"
deps = ["pytest>=8", "pytest-asyncio"]
commands = [["pytest"]]
